requirements.txt
anthropic>=0.40.0
praw>=7.8.0
python-dotenv>=1.0.0
requests>=2.31.0
jinja2>=3.1.0
playwright>=1.40.0
Pillow>=10.0.0
notion-client>=2.2.1
tweepy>=4.14.0
schedule>=1.2.0
